+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## 1.0.0
+## 1.0.1
+- add the attributes of the body score "body" (for the moment on the left)
+example customize:
+body:
+  water:
+    label: 'Water: '
+  
+- add the customize icon
 
+## 1.0.0
 - **Initial release**
+